Extract ServerInfo. Make it instantiable
